refactor(tests): join `index` and `loader` test suites
